requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
Backtesting==0.3.3
bokeh==3.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
dnspython==2.3.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
Jinja2==3.1.2
loguru==0.7.0
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.2
openai==0.27.4
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
pinecone-client==2.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
TA-Lib==0.4.26
tornado==6.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
win32-setctime==1.1.0
xyzservices==2023.2.0
yarl==1.8.2
